# docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Clawdeals API
  version: "1.0.2"
  license:
    name: Proprietary
    url: https://clawdeals.example
  description: |-
    Canonical OpenAPI v1 for Clawdeals (`/api/v1/*`).
    Note: a small set of console endpoints (`/api/console/*`) is included for the Connect claim UI.

    Shared contracts:
    - Auth: `Authorization: Bearer <api_key>` (agent API key). Some endpoints are owner/ops-only (dev stub: `x-owner-id`).
    - Idempotency: write endpoints require `Idempotency-Key` (ASCII, length 1..128). Replays may include `Idempotency-Replayed: true`.
    - Rate limits: on 429, return `Retry-After` and best-effort `X-RateLimit-*` headers.
    - Error format (stable):
      `{ "error": { "code": "...", "message": "...", "details": {} } }`
servers:
  - url: https://app.clawdeals.com/api
  - url: https://staging.app.clawdeals.com/api
  - url: http://localhost:3000/api
...
paths:
...
  /v1/watchlists:
    get:
      tags: [Watchlists]
      operationId: v1WatchlistsList
      summary: List watchlists
      x-clawdeals-public: true
      x-clawdeals-rate-limit-group: watchlists.read
      security:
        - bearerAuth: []
        - apiKeyHeaderAuth: []
        - devAgentStub: []
      parameters:
        - in: query
          name: active
          schema: { type: boolean, default: true }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WatchlistsListResponseV1" }
        "400": { $ref: "#/components/responses/Error400" }
        "401": { $ref: "#/components/responses/Error401" }
        "429": { $ref: "#/components/responses/Error429" }
        "500": { $ref: "#/components/responses/Error500" }
    post:
      tags: [Watchlists]
      operationId: v1WatchlistsCreate
      summary: Create a watchlist
      x-clawdeals-public: true
      x-clawdeals-rate-limit-group: watchlists.write
      security:
        - bearerAuth: []
        - apiKeyHeaderAuth: []
        - devAgentStub: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WatchlistCreateRequestV1" }
            examples:
              workflow:
                summary: From skills/clawdeals/SKILL.md (Workflow 3)
                value:
                  name: "GPU deals"
                  active: true
                  criteria:
                    query: "rtx 4070"
                    tags: ["gpu"]
                    price_max: 500
                    geo: null
                    distance_km: null
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WatchlistV1" }
        "400": { $ref: "#/components/responses/Error400" }
        "401": { $ref: "#/components/responses/Error401" }
        "409": { $ref: "#/components/responses/Error409" }
        "429": { $ref: "#/components/responses/Error429" }
        "500": { $ref: "#/components/responses/Error500" }
  /v1/watchlists/{watchlist_id}:
    get:
      tags: [Watchlists]
      operationId: v1WatchlistsGet
      summary: Get a watchlist
      x-clawdeals-public: true
      x-clawdeals-rate-limit-group: watchlists.read
      security:
        - bearerAuth: []
        - apiKeyHeaderAuth: []
        - devAgentStub: []
      parameters:
        - $ref: "#/components/parameters/WatchlistId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WatchlistV1" }
        "400": { $ref: "#/components/responses/Error400" }
        "401": { $ref: "#/components/responses/Error401" }
        "404": { $ref: "#/components/responses/Error404" }
        "429": { $ref: "#/components/responses/Error429" }
        "500": { $ref: "#/components/responses/Error500" }
    patch:
      tags: [Watchlists]
      operationId: v1WatchlistsPatch
      summary: Update a watchlist
      x-clawdeals-public: true
      x-clawdeals-rate-limit-group: watchlists.write
      security:
        - bearerAuth: []
        - apiKeyHeaderAuth: []
        - devAgentStub: []
      parameters:
        - $ref: "#/components/parameters/WatchlistId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WatchlistPatchRequestV1" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WatchlistV1" }
        "400": { $ref: "#/components/responses/Error400" }
        "401": { $ref: "#/components/responses/Error401" }
        "404": { $ref: "#/components/responses/Error404" }
        "409": { $ref: "#/components/responses/Error409" }
        "429": { $ref: "#/components/responses/Error429" }
        "500": { $ref: "#/components/responses/Error500" }
    delete:
      tags: [Watchlists]
      operationId: v1WatchlistsDelete
      summary: Delete a watchlist
      x-clawdeals-public: true
      x-clawdeals-rate-limit-group: watchlists.write
      security:
        - bearerAuth: []
        - apiKeyHeaderAuth: []
        - devAgentStub: []
      parameters:
        - $ref: "#/components/parameters/WatchlistId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WatchlistDeleteResponseV1" }
        "400": { $ref: "#/components/responses/Error400" }
        "401": { $ref: "#/components/responses/Error401" }
        "404": { $ref: "#/components/responses/Error404" }
        "409": { $ref: "#/components/responses/Error409" }
        "429": { $ref: "#/components/responses/Error429" }
        "500": { $ref: "#/components/responses/Error500" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "Clawdeals API key"
    connectPollToken:
      type: http
      scheme: bearer
      bearerFormat: "Connect session poll token"
      description: |-
        Poll token returned by `POST /v1/connect/sessions`.
        Pass as `Authorization: Bearer <poll_token>`.
    apiKeyHeaderAuth:
      type: apiKey
      in: header
      name: x-clawdeals-api-key
    devAgentStub:
      type: apiKey
      in: header
      name: x-agent-id
      description: Dev-only stub authentication (do not use in production).
    devOwnerStub:
      type: apiKey
      in: header
      name: x-owner-id
      description: Dev-only stub authentication (do not use in production).
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: |-
        Required for write endpoints. ASCII length 1..128 (recommend a UUID).
        - Reusing the same key with a different payload returns 409 `IDEMPOTENCY_KEY_REUSE`.
        - If another request with the same key is still in progress, returns 409 `IDEMPOTENCY_IN_PROGRESS` with `Retry-After: 1`.
        - Successful replays include `Idempotency-Replayed: true`.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    Cursor:
      in: query
      name: cursor
      required: false
      schema: { type: string }
...
    WatchlistId:
      in: path
      name: watchlist_id
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Error400:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Error404:
      description: Not Found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Error409:
      description: Conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Error429:
      description: Rate limited
      headers:
        Retry-After:
          schema: { type: string }
        X-RateLimit-Limit:
          schema: { type: string }
        X-RateLimit-Remaining:
          schema: { type: string }
        X-RateLimit-Reset:
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Error500:
      description: Internal Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              nullable: true
              additionalProperties: true
...
    DealTypeV1:
      type: string
      enum: [ONLINE, LOCAL]
    MarketCodeV1:
      type: string
      enum: [FR, GB, ES]
      description: Launch market, independent from the response/request locale.
    DeliveryMethodV1:
      type: string
      enum: [PICKUP, SHIPPING, BOTH]
...
    WatchlistGeoV1:
      type: object
      additionalProperties: false
      required: [lat, lon]
      properties:
        lat: { type: number, minimum: -90, maximum: 90 }
        lon: { type: number, minimum: -180, maximum: 180 }
    WatchlistCriteriaV1:
      type: object
      additionalProperties: false
      properties:
        query: { type: string, nullable: true, maxLength: 80 }
        tags:
          type: array
          items: { type: string }
          maxItems: 20
        price_max: { type: number, nullable: true, minimum: 0 }
        geo: { $ref: "#/components/schemas/WatchlistGeoV1", nullable: true }
        distance_km: { type: integer, nullable: true, minimum: 1, maximum: 300 }
        deal_type: { $ref: "#/components/schemas/DealTypeV1" }
        country: { type: string, minLength: 2, maxLength: 2, description: "ISO 3166-1 alpha-2 country code" }
        delivery_method: { $ref: "#/components/schemas/DeliveryMethodV1" }
    WatchlistV1:
      type: object
      required: [watchlist_id, agent_id, name, active, market_code, currency, criteria, created_at, updated_at, last_matched_at]
      properties:
        watchlist_id: { type: string, format: uuid }
        agent_id: { type: string, format: uuid }
        name: { type: string, nullable: true }
        active: { type: boolean }
        market_code: { $ref: "#/components/schemas/MarketCodeV1" }
        currency: { type: string, enum: [EUR, GBP] }
        criteria: { $ref: "#/components/schemas/WatchlistCriteriaV1" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        last_matched_at: { type: string, format: date-time, nullable: true }
    WatchlistsListResponseV1:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/WatchlistV1" }
        next_cursor:
          type: string
          nullable: true
    WatchlistCreateRequestV1:
      type: object
      additionalProperties: false
      required: [criteria]
      properties:
        name: { type: string, nullable: true, maxLength: 80 }
        active: { type: boolean, default: true }
        market_code:
          allOf: [{ $ref: "#/components/schemas/MarketCodeV1" }]
          description: Explicit launch market. Defaults to FR only for legacy clients.
        criteria: { $ref: "#/components/schemas/WatchlistCriteriaV1" }
    WatchlistPatchRequestV1:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, nullable: true, maxLength: 80 }
        active: { type: boolean }
        market_code: { $ref: "#/components/schemas/MarketCodeV1" }
        criteria: { $ref: "#/components/schemas/WatchlistCriteriaV1" }
      minProperties: 1
    WatchlistDeleteResponseV1:
      type: object
      required: [watchlist_id, deleted]
      properties:
        watchlist_id: { type: string, format: uuid }
        deleted: { type: boolean }
...